redo bank, drop account_balance table and store directly to top_accounts table
