+++ database/bank.go
@@ -4,8 +4,6 @@ import (
 	"fmt"
 
 	dbtypes "github.com/forbole/bdjuno/v3/database/types"
-	dbutils "github.com/forbole/bdjuno/v3/database/utils"
-	"github.com/forbole/bdjuno/v3/types"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/lib/pq"
@@ -26,50 +24,5 @@ WHERE supply.height <= excluded.height`
 		return fmt.Errorf("error while storing supply: %s", err)
 	}
 
-	return nil
-}
-
-// SaveAccountBalances allows to store the given balances inside the database
-func (db *Db) SaveAccountBalances(balances []types.NativeTokenBalance) error {
-	paramsNumber := 3
-	slices := dbutils.SplitBalances(balances, paramsNumber)
-
-	for _, balances := range slices {
-		if len(balances) == 0 {
-			continue
-		}
-
-		// Store up-to-date data
-		err := db.saveUpToDateBalances(paramsNumber, balances)
-		if err != nil {
-			return fmt.Errorf("error while storing up-to-date balances: %s", err)
-		}
-	}
-
-	return nil
-}
-
-func (db *Db) saveUpToDateBalances(paramsNumber int, balances []types.NativeTokenBalance) error {
-	stmt := `INSERT INTO account_balance (address, coin, height) VALUES `
-	var params []interface{}
-
-	for i, bal := range balances {
-		bi := i * paramsNumber
-		stmt += fmt.Sprintf("($%d, $%d, $%d),", bi+1, bi+2, bi+3)
-
-		params = append(params, bal.Address, bal.Balance.String(), bal.Height)
-	}
-
-	stmt = stmt[:len(stmt)-1]
-	stmt += `
-ON CONFLICT (address) DO UPDATE 
-	SET coin = excluded.coin, 
-	    height = excluded.height 
-WHERE account_balance.height <= excluded.height`
-
-	_, err := db.Sql.Exec(stmt, params...)
-	if err != nil {
-		return err
-	}
 	return nil
 }
