+++ database/bank_test.go
@@ -4,7 +4,6 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
 	dbtypes "github.com/forbole/bdjuno/v3/database/types"
-	"github.com/forbole/bdjuno/v3/types"
 )
 
 func (suite *DbTestSuite) TestBigDipperDb_SaveSupply() {
@@ -73,106 +72,4 @@ func (suite *DbTestSuite) TestBigDipperDb_SaveSupply() {
 	suite.Require().NoError(err)
 	suite.Require().Len(rows, 1, "supply table should contain only one row")
 	suite.Require().True(expected.Equals(rows[0]))
-}
-
-func (suite *DbTestSuite) TestSaveAccountBalance() {
-	address1 := suite.getAccount("cosmos140xsjjg6pwkjp0xjz8zru7ytha60l5aee9nlf7")
-	address2 := suite.getAccount("cosmos1tcpsdy9alvucwj0h23n56tey6zmrvkm7sndh9j")
-
-	// Save the data
-	err := suite.database.SaveAccountBalances([]types.AccountBalance{
-		types.NewAccountBalance(
-			address1.String(),
-			sdk.NewCoins(
-				sdk.NewCoin("desmos", sdk.NewInt(10)),
-				sdk.NewCoin("uatom", sdk.NewInt(20)),
-			),
-			10,
-		),
-		types.NewAccountBalance(
-			address2.String(),
-			sdk.NewCoins(
-				sdk.NewCoin("uatom", sdk.NewInt(100)),
-			),
-			10,
-		),
-	})
-	suite.Require().NoError(err)
-
-	// Verify the data
-	expected := []dbtypes.AccountBalanceRow{
-		dbtypes.NewAccountBalanceRow(
-			address1.String(),
-			dbtypes.NewDbCoins(sdk.NewCoins(
-				sdk.NewCoin("desmos", sdk.NewInt(10)),
-				sdk.NewCoin("uatom", sdk.NewInt(20)),
-			)),
-			10,
-		),
-		dbtypes.NewAccountBalanceRow(
-			address2.String(),
-			dbtypes.NewDbCoins(sdk.NewCoins(
-				sdk.NewCoin("uatom", sdk.NewInt(100)),
-			)),
-			10,
-		),
-	}
-
-	var rows []dbtypes.AccountBalanceRow
-	err = suite.database.Sqlx.Select(&rows, `SELECT * FROM account_balance ORDER BY address`)
-	suite.Require().NoError(err)
-	suite.Require().Len(rows, len(expected))
-
-	for index, row := range rows {
-		suite.Require().True(row.Equal(expected[index]))
-	}
-
-	// Update the data
-	err = suite.database.SaveAccountBalances([]types.AccountBalance{
-		types.NewAccountBalance(
-			address1.String(),
-			sdk.NewCoins(
-				sdk.NewCoin("desmos", sdk.NewInt(10)),
-			),
-			9,
-		),
-		types.NewAccountBalance(
-			address2.String(),
-			sdk.NewCoins(
-				sdk.NewCoin("uatom", sdk.NewInt(100)),
-				sdk.NewCoin("desmos", sdk.NewInt(200)),
-			),
-			11,
-		),
-	})
-	suite.Require().NoError(err)
-
-	// Verify the data
-	expected = []dbtypes.AccountBalanceRow{
-		dbtypes.NewAccountBalanceRow(
-			address1.String(),
-			dbtypes.NewDbCoins(sdk.NewCoins(
-				sdk.NewCoin("desmos", sdk.NewInt(10)),
-				sdk.NewCoin("uatom", sdk.NewInt(20)),
-			)),
-			10,
-		),
-		dbtypes.NewAccountBalanceRow(
-			address2.String(),
-			dbtypes.NewDbCoins(sdk.NewCoins(
-				sdk.NewCoin("uatom", sdk.NewInt(100)),
-				sdk.NewCoin("desmos", sdk.NewInt(200)),
-			)),
-			11,
-		),
-	}
-
-	rows = []dbtypes.AccountBalanceRow{}
-	err = suite.database.Sqlx.Select(&rows, `SELECT * FROM account_balance ORDER BY address`)
-	suite.Require().NoError(err)
-	suite.Require().Len(rows, len(expected))
-
-	for index, row := range rows {
-		suite.Require().True(row.Equal(expected[index]))
-	}
 }
