+++ modules/bank/utils_update_balances.go
@@ -9,18 +9,35 @@ import (
 
 // UpdateBalances updates the balances of the accounts having the given addresses,
 // taking the data at the provided height
-func (m *Module) UpdateBalances(addresses []string, height int64) ([]types.NativeTokenBalance, error) {
+func (m *Module) UpdateBalances(addresses []string, height int64) error {
 	log.Debug().Str("module", "bank").Int64("height", height).Msg("updating balances")
 
 	balances, err := m.keeper.GetBalances(addresses, height)
 	if err != nil {
-		return nil, fmt.Errorf("error while getting account balances: %s", err)
+		return fmt.Errorf("error while getting account balances: %s", err)
 	}
 
-	err = m.db.SaveAccountBalances(balances)
+	err = m.saveTopAccountsAvailable(balances, height)
 	if err != nil {
-		return nil, fmt.Errorf("error while saving account balances: %s", err)
+		return fmt.Errorf("error while saving top accounts available balance: %s", err)
 	}
 
-	return balances, nil
+	return nil
+}
+
+func (m *Module) saveTopAccountsAvailable(balances []types.NativeTokenBalance, height int64) error {
+	if len(balances) == 0 {
+		return nil
+	}
+	err := m.db.SaveTopAccountsBalance("available", balances)
+	if err != nil {
+		return fmt.Errorf("error while saving top accounts available balances: %s", err)
+	}
+
+	var addresses = make([]string, len(balances))
+	for index, bal := range balances {
+		addresses[index] = bal.Address
+	}
+
+	return nil
 }
